requirements.txt
pymupdf==1.23.20
openai==0.28
python-dotenv==1.0.1
